backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
python-dotenv
psycopg2-binary
